Exclude confidential tests ga4gh/dockstore#1068 (#1070)
Add ConfidentialTest annotation to all tests that have a dependency
on confidential data. CommonTestUtilities needs dockstoreTest.yml
to be present, and that is only present with confidential data.

I did try alternatively serving up a different .yml file, but the tests
would then fail because of requirements for other confidential data, so
just added the annotation to all tests that require dockstoreTest.yml.