Change /workflows/sharedworkflow response entity
Now returns a List<SharedWorkflows> instead of List<Workflow>, where
SharedWorkflows class has a Role and a list of workflows.

Initially attempted to have the endpoint return a
Map<Role, List<Workflow>>, but the problem when I do that is that
the generated Swagger Java client returns a Map<Role, Object>.

This response is slightly more cumbersome, but it preserves the types
in the Swagger clients, so it seems better.