For shared flows, don't show roles without workflows (#1567)
If a SharedWorkflows object has no workflows, don't return it.

This can happen if workflowDAO.findByPath doesn't find the any of the
workflows for a given role. I believe this is an unlikely scenario, but
I've seen it an a dev environment where I register a resource with SAM,
then effectively delete the resource by reinitializing my database, so
that SAM ends up pointing to a resource that doesn't exist.