Accept Google tokens from users not in Dockstore (#1566)
* Accept Google tokens from users not in Dockstore

If API call made with a Google token not issued against the Dockstore
audience (client id), accept it if it's issued against a whitelisted
Google client id (see
DockstoreWebserviceConfiguration.externalGoogleClientIds).

If the user is not in Dockstore, create the user. The main difference
for a User created this way vs. a user signing in is that there is no
refresh token.

Added org.mockito.plugins.MockMaker file so I could mock Userinfoplus,
a final class, which Mockito doesn't let you mock by default.