Don't Return Duplicate Shared Workflows
#1609

In SAM, for the same resource id, the same email can be added to
more than one policy for that resource.

That would lead to the same workflow showing up more than once, each
time for a different role, on the shared flows endpoint.

Now weed out duplicates by only returning the resource of the most
privileged role.