Correct Check for Original Owner
#1657

Problem was that I was comparing email with User.username, which
will not be the Google email, if the user first
signed in with GitHub. Now also check profile for an email in there.

Also:

* Add the check when adding a permission; originally only checked
when removing a permission. Basically we shouldn't let you muck
with an original owner's permissions at all.