Add Data Object Service File plugin
Added release 0.0.1 to plugins.json.

Download didn't work, because... HEAD call to

https://github.com/dockstore/data-object-service-plugin/releases/
download/0.0.1/dockstore-file-dos-plugin-0.0.1.zip

was failing(!).

This is because when calling srcContent.getSize(), it would ultimately
call org.apache.commons.vfs2.provider.http.HttpFileObject.doGetType,
which would do a HEAD, and the GitHub API returns a 403.

So wrapped call to get the size in a try...catch. If the size
doesn't come back, then don't display a progress bar.

I believe this could fix hypothetical cases where there is no
size in the header, e.g, if there is a chunked response.