Properly look up user for calls with Google token
#1684

When web service gets a call with a Google token, look for an existing
user by searching User.Profile for that email address.

In progress, because:

1. The query is very inefficient. I think we would want indexes
on token_type and email columns in user_profile. I need to figure out
how to do that with Liquibase, but I don't want to pursue until we
get agreement on the next item, or some different approach.
2. If we allow for multiple Dockstore accounts pointing to the same
Google account then there will be no way to have this work reliably,
as there would be no way to know which Dockstore account to
associate the call with.