When self-destructing, remove sharing records (#1733)
#1719 

Add `isSharing` to `PermissionsInterface`. Call it to check if a user can delete
herself/himself -- in addition to any other rules, users should not be able to delete
themselves if they are sharing workflow(s) with others.

Add `selfDestruct` to `PermissionsInterface`, which will delete any sharing records. For SAM, this means that we remove the resourceIds that the user owns.