Fix owner losing permissions to their own SAM resource
#1805

Problem was that a new owner policy was being created when adding
another owner, and the new owner policy only added the new owner.

1. We don't need to ever create an owner policy explicitly, as SAM
is configured for dockstore-tool resource to always create a default
owner policy. This is unlike the writer and reader policies, which we do
need to explicitly create.
2. The problem originated in that I was passing the list of policies
the new user already belonged to ensurePolicyExists(), which would
check to see if the owner policy existed, and it wasn't in the list.
3. Two fixes for that, to be safe.
    1. Don't ever let ensurePolicyExists() create an owner policy
    2. Pass all of a resource's existing policies to ensurePolicyExists,
    instead of just the ones the new user already belongs to.
4. In doing that, I noticed that initializePermissions was not
returning the policy's permissions.
5. Finally, in looking at this, I noticed the setPermission already
returns the list of permissions, and WorkflowResource was
unnecessarily fetching them again.