Don't log error when saving a non-published workflow
#2034

Saving a new hosted workflow version calls
`ElasticManager.handleIndexUpdate`, which would cause an error to be
logged if the the workflow is not published. Log it at info level
instead, since it's not an error condition.

The other option is to not call the method at all if the workflow is
not published, but the method is called from several places, and that
seemed slightly riskier.

Also noticed that if there is an error in the response from the
ElasticManager, that wasn't being logged at the error level.