Remove maven release steps from README
Doing a maven release is described in two places, here, and in
internal documentation.

Instead of maintaining the same info in two places and there being
confusion as to which one to follow, I propose we
just maintain the internal doc up to date and remove this info from
the README.

We can do it the other way around, if preferred. Please reject the
associated PR if so.