Fix DAG for HCA SnapAtac workflow
#2805

For this workflow, one inputMap._2 didn't have a
head property but instead had a tail.

Filtered the tail out with pattern matching.

A graph now does display and it looks correct.

This seems to be be caused by variables in tasks, e.g.,
Float input_size = size(input_bam, "GiB"); commenting out
all instances of variables in the tasks causes the
original error to go away.