Switch to postgres 11.6
#2321

11.6 instead of 11.7 because there is no 11.7-alpine-ram
on CircleCI, although I don't know if we need the alpine
version.

By default, Postgres 11 does not trust local connections,
and needs to be so configured.

It was easier to run PG as a Docker container and take
advantage of the POSTGRES_HOST_AUTH_METHOD env
variable, rather than modifying the conf file for the one installed via addon and then restarting it.