feature/3712/lazyLoadSourcefiles (#3740)
#3712
Sourcefiles will get loaded on an entry during refresh. There are some calls to getSourcefile() so they have to be ejected out before returning. Sourcefiles are also used during refresh to update the verified information in Version.java.

But just getting the workflow/tool when looking at one on Dockstore shouldn't load them in the webservice.