Set up git-secrets/maven integration (#3788)
* Add an exec plugin that sets git config during initialize
Set up git secrets AWS provider
Make sure we don't set extra configs

* Move installation to a bash script

* Clear false positives in current source

* Add readme, comments to .gitallowed

* Move script, simplify, set bash flags

* Install git-secrets in CI

* Change gitallowed to be more paranoid

* Run git-secrets scan in circle
Technically if someone pushes a secret to a branch
it will still be in the git tree, however, this
check will help it keep from being merged.

Fix travis wget script
And add git-secrets to lgtm

* Clean up .gitallowed

* One lgtm.yml to rule them all

* Accepted suggestions by @chmreid

* Use preferred hidden lgtm yml

* Travis failures need a new commit
Something is failing when a build is rerun and a coveralls
support has already been submitted. Trying a workaround of
just making a new commit to rerun the tests. This will rerun
all the tests. Sorry :/