Detect if WDL is using parameterized Docker images (#4031)
https://ucsc-cgl.atlassian.net/browse/SEAB-1971

Figures out if a reference to a Docker image in WDL is a parameter or a literal string. Doesn't do anything with that info yet.

Changed a Triple<String, String, String> to DockerInfo because I was having a hard time keeping track of which which tuple value was which, and because there is now a fourth value.