Signal Retry to Lambda on GitHub Rate Limit failure (#4134)
#3770

Use a variant of RateLimitHandler.FAIL to cause failure, then throw 500 error for Lambda.

In looking at this, I now consider the RateLimitHandler.WAIT we used to use as a possible culprit for our most recent prod incident. The default WAIT handler will wait until the GitHub rate limit resets, which can be be lots of minutes, since the rate limit resets every hour. I couldn't find any proof of it, but it would block a thread and have an open DB transaction, so it does seem potentially dangerous.

Regarding other points on the ticket, a ConstraintViolation already signals retry -- it happens when committing the transaction fails, and that gets translated to a 5xx error.