Upgrade Hibernate and run unit tests with PG 11 (#4167)
https://ucsc-cgl.atlassian.net/browse/SEAB-2676,

Change unit tests to use PG 11, which opened a whole can of worms. Lots of details in SEAB-2676. Short version here:

Postgres 10 introduced support for SQL Identity.

Pre PG-10, Liquibase/Hibernate would simulate SQL Identity with SQL Sequence.

When we upgraded from PG 9 to PG11, Liquibase took advantage of SQL Identity, but Hibernate didn't, leading to the propose_migrations failures (Liquibase would generate identity, Hibernate would generate Sequence)

We need to tell Hibernate to use the PG10 dialect in the dockstore.yml, so that it will also generate Identity, but our version of Hibernate doesn't have that dialect.

So, upgrade Hibernate to a version that has the PostgreSQL10 dialect. The version bundled with DropWizard (we were behind), supports it. So upgrading it just means not pinning to an older version. And then chasing down the enforcer plugin errors.

Also:

Run CWL tests in CircleCI, remove Travis

And, saw some OOM errors, which removing the unused scala plugin seemed to fix.