Prevent duplicate paths between apptools and workflows (#4655)
* Prevent duplicate paths between apptools and workflows

Don't allow there to be duplicate paths between apptools and workflows. This PR creates a master list of all workflow/apptool paths in a new table called fullworkflowpath and there is a new trigger that will insert a new row into this table every time a row is inserted into the either the workflow or apptool tables. Will do something similar to deletions as well

https://ucsc-cgl.atlassian.net/browse/SEAB-3734

Co-authored-by: Gary Luu <garylau.work@gmail.com>