handle null size in Quay Tag (#4722)
Description
The error was being caused by multi-arch images being pushed to Quay. This PR let's the code continue without failing and fixes the test that was being ignored, but it's not quite correct.

Dockstore currently supports multi-arch images for DockerHub and saves this info as a list of Images for each Version. The DockerHub API makes it clear that the different images actually belong to the same version. However, the Quay API isn't as clear...So PR will make separate tags with one image each (where the image can have the size and imageid be null)

I think it's possible to store it in the same way, but it seems complicated looking through their API.

Issue
#4663