Check open data when saving hosted workflow version.
#5339

We had only been checking open data on source control updates, which is
why it worked for source control backed workflows and not hosted
workflows.

In adding a test to CheckUrlHelperFullIT, I discovered that the test
file wasn't running at all on CircleCI. Turned out to be because the
file hadn't been completely migrated to JUnit 5. Updated it to JUnit 5,
and also switched it to use OpenAPI instead of Swagger, as well as
adding a test for the hosted workflow bug.